% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annulus.R
\name{annulus}
\alias{annulus}
\title{Annulus}
\usage{
annulus(im, R, dR = 2, center = c(nrow(im)/2, ncol(im)/2))
}
\arguments{
\item{im}{matrix.}

\item{R}{radius of annulus.}

\item{dR}{annulus width in pixels (default is 2 px).}

\item{center}{point of center c(x, y) (default is center of matrix c(nrow(im) / 2, ncol(im) / 2)).}
}
\value{
Vector with points of annulus.
}
\description{
Get points from annulus on matrix
}
\examples{
m <- matrix(1:512, nrow = 512, ncol = 512)
annulus_points <- annulus(m, R = 57, dR = 1, center = c(100, 100))
}
