% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/besselian_year.R
\name{besselian_year}
\alias{besselian_year}
\title{Besselian year}
\usage{
besselian_year(date, time)
}
\arguments{
\item{date}{a character string (dd-Mon-year)}

\item{time}{a character string (hh:mm:ss)}
}
\value{
Besselian year.
}
\description{
Returns besselian year calculated from date and time
}
\examples{
besselian_year('12-Dec-2009', '02:53:10')
}
