% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bta_list.R
\name{bta_list}
\alias{bta_list}
\title{Prepare data for observations on BTA
with bta_list program}
\usage{
bta_list(objects = NULL)
}
\arguments{
\item{objects}{a character vector with objects}
}
\value{
Tibble with data formatted for bta_list.
}
\description{
Returns tibble with objects formatted for bta_list
}
\examples{
bta_list_data <- bta_list(c('HD 6757', 'HIP 11569'))
}
