% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_functions.R
\name{n_frames}
\alias{n_frames}
\title{TTT}
\usage{
n_frames(n, start = 1)
}
\arguments{
\item{n}{A number.}

\item{start}{A number.}
}
\value{
The array of \code{n} frames.
}
\description{
Take N frames from series of speckle images 512x512x2(bytes)
}
\examples{
n_frames(50) # take first 50 frames
n_frames(15, 10) or
}
