% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obs_database.R
\name{obs_database}
\alias{obs_database}
\title{Observational journals}
\usage{
obs_database(logs_dir = NULL)
}
\arguments{
\item{logs_dir}{a character string of path to folder with observational journals}
}
\value{
Data frame of observational journals.
}
\description{
Returns data frame of BTA observational journals
}
\examples{
db <- obs_database()
}
