% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/par_angle.R
\name{par_angle}
\alias{par_angle}
\title{Calculate parallactic angle}
\usage{
par_angle(log_file = file.choose())
}
\arguments{
\item{log_file}{a character string with the path name to a log file.}
}
\value{
Tibble with names and parallactic angles calculated from BTA log file.
}
\description{
Calculate parallactic angle
}
\examples{
log_filename <- system.file("extdata", "2010-02-27.tel", package = "specklestar")
par_angle(log_filename)
}
