% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/par_angle_BTA.R
\name{par_angle_BTA}
\alias{par_angle_BTA}
\title{Calculate parallactic angles from Alpha, Delta and Stime for BTA location}
\usage{
par_angle_BTA(Alpha = NULL, Delta = NULL, Stime = NULL)
}
\arguments{
\item{Alpha}{a character string (hh:mm:ss)}

\item{Delta}{a character string (dd:mm:ss)}

\item{Stime}{a character string (hh:mm:ss)}
}
\value{
Parallactic angle in degrees.
}
\description{
Calculate parallactic angles from Alpha, Delta and Stime for BTA location
}
\examples{
par_angle_BTA(Alpha = '00:33:32.7', Delta = '55:18:59.5', Stime = '06:53:00')
par_angle_BTA()
}
