% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/par_angle_log.R
\name{par_angle_log}
\alias{par_angle_log}
\title{Parallactic angles}
\usage{
par_angle_log(log_file = NULL, log_date = NULL)
}
\arguments{
\item{log_file}{a character string with the path name to a log file.}

\item{log_date}{a character string with date of observations (YYYY-MM-DD).}
}
\value{
data.frame with names, parallactic angles (in degrees) and mean date
and UTC time of observations calculated from BTA log file.
}
\description{
Returns data.frame with parallactic angles calculated from BTA log file
}
\examples{
log_filename <- system.file("extdata", "2010-02-27.tel", package = "specklestar")
par_angle_log(log_filename, '2010-02-27')
}
