% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/par_angle_manual.R
\name{par_angle_manual}
\alias{par_angle_manual}
\title{Calculate parallactic angles from manually entered data}
\usage{
par_angle_manual()
}
\value{
Parallactic angle in degrees from manually entered data.
}
\description{
Calculate parallactic angles from manually entered data
}
\examples{
par_angle_manual()
}
