% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ps_diff}
\alias{ps_diff}
\title{Power spectrum calculation}
\usage{
ps_diff(filename, threshold = 50000L)
}
\arguments{
\item{filename}{A string.}
}
\value{
The 513 x 1024 double vector of power spectrum.
}
\description{
Power spectrum of the difference of neighboring frames
in the series of speckle images
}
\examples{
pow_spec_diff <- ps_diff(file.choose())

# Plot
library(imager)
plot(as.cimg(pow_spec^0.01))
}
