% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp_code.R
\name{sp_code}
\alias{sp_code}
\title{Spectral code}
\usage{
sp_code(spectr)
}
\arguments{
\item{spectr}{a character string with spectr type of star.}
}
\value{
A number.
}
\description{
Return a number based on spectral type of star
(first two characters)
}
\examples{
spectr_code <- sp_code('B9IVpSi')
}
