% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_functions.R
\name{spe2dat}
\alias{spe2dat}
\title{SPE to dat converter}
\usage{
spe2dat(SPE_file = file.choose(), dat_file)
}
\arguments{
\item{SPE_file}{a character string with the path name to SPE file.}

\item{dat_file}{a character string with the path name to output dat file.}
}
\value{
Saved converted dat file on the disk.
}
\description{
Convert SPE to dat (i.e. removing 4100 byte SPE header) and save it on disk
}
\examples{
\dontrun{
# On Unix-like operating systems only
# Convert choosed SPE file to dat and save it as your_file.dat
spe2dat(, "your_file.dat")
}
}
