% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/speckle_binary.R
\name{speckle_binary}
\alias{speckle_binary}
\title{Calculation rho, theta and dm for binary star}
\usage{
speckle_binary(object_file = file.choose(), dark_file = file.choose(),
  flat_file = file.choose())
}
\arguments{
\item{object_file}{A string.}

\item{dark_file}{A string.}

\item{flat_file}{A string.}
}
\value{
double vector of rho, theta and dm.
}
\description{
Obtaining positional parameters and magnitude
differenece between components of binary star
from the series of speckle images
}
\examples{
# parameters <- speckle_binary()
}
