% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{speckle_ps_diff}
\alias{speckle_ps_diff}
\title{Power spectrum calculation}
\usage{
speckle_ps_diff(filename, threshold = 50000L)
}
\arguments{
\item{filename}{A string.}

\item{threshold}{An integer (default 50000).}
}
\value{
The 513 x 1024 double vector of power spectrum.
}
\description{
Power spectrum of the difference of neighboring frames
in the series of speckle images
}
\examples{
obj_filename <- system.file("extdata", "ads15182_550_5_frames.dat", package = "specklestar")
pow_spec_diff <- speckle_ps_diff(obj_filename)
}
