% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{speckle_stat}
\alias{speckle_stat}
\title{Statistics of speckles}
\usage{
speckle_stat(filename)
}
\arguments{
\item{filename}{A string.}
}
\value{
The double vector of speckle statistics.
}
\description{
Calculate statistics of speckles in the series of speckle images and filter "bad" frames
}
\examples{
spec_stat <- speckle_stat(file.choose())

# Plot
plot(speckle_stat)
}
