% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specklestar.R
\docType{package}
\name{specklestar}
\alias{specklestar}
\alias{specklestar-package}
\title{specklestar: A package for reduction of speckle data.}
\description{
The specklestar package provides three categories of important functions:
foo, bar and baz.
}
\section{Foo functions}{

The foo functions ...
}

