% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_functions.R
\name{spool_assemble}
\alias{spool_assemble}
\title{Assemble spool of speckle images}
\usage{
spool_assemble()
}
\value{
The .dat file.
}
\description{
Convert spool of speckle images to single .dat file
}
\examples{
# spool_assemble(50)
}
