% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vizier_query.R
\name{vizier_query}
\alias{vizier_query}
\title{VizieR query}
\usage{
vizier_query(vizier_table = NULL)
}
\arguments{
\item{vizier_table}{a character string with name of VizieR table.}
}
\value{
Tibble with data.
}
\description{
Returns tibble of selected table from VizieR
}
\examples{
vizier_data <- vizier_query('J/other/AstBu/63.278/table1')
}
