% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vizier_query_param.R
\name{vizier_query_param}
\alias{vizier_query_param}
\title{VizieR query by constrain on parameter}
\usage{
vizier_query_param(vizier_table = NULL, par = NULL, constrain = NULL)
}
\arguments{
\item{vizier_table}{a character string with name of VizieR table.}

\item{constrain}{a character string with constrain.}

\item{param}{a character string with parameter.}
}
\value{
Tibble with data.
}
\description{
Returns tibble of selected table from VizieR
}
\details{
http://vizier.u-strasbg.fr/vizier/doc/asu-summary.htx

https://vizier.u-strasbg.fr/vizier/vizHelp/args.htx
}
\examples{
vizier_data <- vizier_query_param('J/other/AstBu/63.278/table1', 'dm', '<2')
vizier_wds <- vizier_query_param('J/other/AstBu/63.278/table1', 'Name', c('G76-21', 'G63-46'))
}
