% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wds_id.R
\name{wds_id}
\alias{wds_id}
\title{WDS index}
\usage{
wds_id(ra, dec, delim = " ", name = "")
}
\arguments{
\item{ra}{a character string with right ascension.}

\item{dec}{a character string with declination.}

\item{delim}{a character string with delimiter used in both ra and dec.}

\item{name}{a character string with extra name.}
}
\value{
A number.
}
\description{
Return WDS index based on right ascension and declination
}
\examples{
wds_indx1 <- wds_id('00 00 43.63', '+45 15 12.0')
wds_indx2 <- wds_id('16:37:44.07', '+19:56:55.1', ':', 'wds')
}
