% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bta_list.R
\name{bta_list}
\alias{bta_list}
\title{Prepare data for observations on BTA
with bta_list program}
\usage{
bta_list(objects = NULL, epoch = "01-Jan-2000")
}
\arguments{
\item{objects}{a character vector with objects.}

\item{epoch}{a character with epoch (default is '01-Jan-2000').}
}
\value{
data.frame with data formatted for bta_list.
}
\description{
Returns data.frame with objects formatted for bta_list
}
\examples{
bta_list_data <- bta_list(c('HD 6757', 'HIP 11569'))
}
