% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{middle_frame}
\alias{middle_frame}
\title{Middle frame}
\usage{
middle_frame(filename, subtrahend, threshold = 50000L)
}
\arguments{
\item{filename}{A string.}

\item{subtrahend}{512 x 512 matrix to subtract.}

\item{threshold}{An integer (default 50000).}
}
\value{
The 512 x 512 matrix of middle speckle image.
}
\description{
Average image of the series of 512 x 512 px images
}
\examples{
obj_filename <- system.file("extdata", "ads15182_550_2_frames.dat", package = "specklestar")
zero_matrix <- matrix(0, 512, 512)
mf <- middle_frame(obj_filename, subtrahend = zero_matrix)
}
