% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{middle_frame}
\alias{middle_frame}
\title{Middle frame}
\usage{
middle_frame(filename)
}
\arguments{
\item{filename}{A string.}
}
\value{
The array of middle speckle image.
}
\description{
Average frame of series of speckle images
}
\examples{
library(imager)
mf <- middle_frame(filename)
plot(as.cimg(mf))
}
