% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{middle_frame}
\alias{middle_frame}
\title{Middle frame}
\usage{
middle_frame(filename)
}
\arguments{
\item{filename}{A string.}
}
\value{
The vector of middle speckle image.
}
\description{
Average frame of series of speckle images
}
\examples{
mf <- middle_frame(file.choose())
mf <- matrix(mf, 512, 512)

# Plot
library(imager)
plot(as.cimg(mf))
}
