% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{middle_frame}
\alias{middle_frame}
\title{Middle frame}
\usage{
middle_frame(filename, subtrahend, threshold = 50000L)
}
\arguments{
\item{filename}{A string.}
}
\value{
The 512 x 512 matrix of middle speckle image.
}
\description{
Average image of the series of 512 x 512 px images
}
\examples{
mf <- middle_frame(file.choose())

# Plot
library(imager)
plot(as.cimg(mf))
}
