% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_functions.R
\name{n_frames}
\alias{n_frames}
\title{N frames from file}
\usage{
n_frames(input_file = file.choose(), output_file, n, from = 1)
}
\arguments{
\item{input_file}{a character string with the path name to input dat file.}

\item{output_file}{a character string with the path name to output dat file.}

\item{n}{a number of frames.}

\item{from}{a number of first frame.}
}
\value{
The array of \code{n} frames.
}
\description{
Take N frames from series of 512 x 512 speckle images
and write them to separate binary file
}
\examples{
\dontrun{
# On Unix-like operating systems only
# Write first 50 frames from choosed file to 50_frames.dat file
n_frames(, "50_frames.dat", 50)
}
}
