% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ps}
\alias{ps}
\title{Power Spectrum calculation}
\usage{
ps(filename, threshold = 50000L)
}
\arguments{
\item{filename}{A string.}
}
\value{
The 512 x 512 double vector of Power Spectrum.
}
\description{
Power Spectrum of the series of speckle images
}
\examples{
pow_spec <- ps(file.choose())

# Plot
library(imager)
plot(as.cimg(pow_spec^0.01))
}
