% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ps}
\alias{ps}
\title{Power spectrum calculation}
\usage{
ps(filename, dark, flat, threshold = 50000L)
}
\arguments{
\item{filename}{A string.}

\item{512}{x 512 middle frame matrix}

\item{512}{x 512 middle flat field matrix}
}
\value{
The 513 x 1024 double vector of power spectrum.
}
\description{
Power spectrum of the series of 512 x 512 speckle images
}
\examples{
# Suppose we have midd_dark and midd_flat 512 x 512 matrices
pow_spec <- ps(file.choose(), dark = midd_dark, flat = midd_flat)

# Plot
library(imageviewer)
imageviewer(log10(pow_spec))
}
