% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_functions.R
\name{spe2dat}
\alias{spe2dat}
\title{SPE to dat converter}
\usage{
spe2dat(SPE_file)
}
\arguments{
\item{SPE_file}{A string.}
}
\value{
Save converted dat file on the disk in the same folder.
}
\description{
Convert SPE to dat (i.e. removing 4100 byte SPE header)
}
\examples{
# spe2dat(file.choose())
}
