% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{speckle_acf}
\alias{speckle_acf}
\title{Autocorrelation function calculation}
\usage{
speckle_acf(ps)
}
\arguments{
\item{ps}{513 x 1024 power spectrum double matrix.}
}
\value{
The 513 x 1024 double matrix of ACF.
}
\description{
Autocorrelation function of power spectrum
}
\examples{
obj_filename <- system.file("extdata", "ads15182_550_2_frames.dat", package = "specklestar")
pow_spec_diff <- speckle_ps_diff(obj_filename)
acf <- speckle_acf(pow_spec_diff)
}
