% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/speckle_frame.R
\name{speckle_frame}
\alias{speckle_frame}
\title{Get selected speckle frame
Get specified speckle frame as matrix from file}
\usage{
speckle_frame(data_file = file.choose(), frame = 1)
}
\arguments{
\item{data_file}{a character string with the path name to a file.}

\item{frame}{an integer.}
}
\value{
512 x 512 matrix with given frame.
}
\description{
Get selected speckle frame
Get specified speckle frame as matrix from file
}
\examples{
\dontrun{
# On Unix-like operating systems only
# Read frame number 2 from file to matrix
obj_filename <- system.file("extdata", "ads15182_550_2_frames.dat", package = "specklestar")
frame2 <- speckle_frame(obj_filename, 2)
}
}
