% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/speckle_frame.R
\name{speckle_frame}
\alias{speckle_frame}
\title{Get selected speckle frame}
\usage{
speckle_frame(data_file = file.choose(), frame = 1)
}
\arguments{
\item{data_file}{A string.}

\item{frame}{An integer.}
}
\value{
512 x 512 matrix with given frame
}
\description{
Get specified speckle frame as matrix from file
}
\examples{
# Read frame number 3 from file
obj_filename <- system.file("extdata", "ads15182_550_5_frames.dat", package = "specklestar")
frame3 <- speckle_frame(obj_filename, 3)
}
