% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/speckle_frame.R
\name{speckle_frame}
\alias{speckle_frame}
\title{Plot selected speckle frame}
\usage{
speckle_frame(data_file = file.choose(), frame = 1)
}
\arguments{
\item{filename}{A string.}
}
\value{
NULL.
}
\description{
Ploting specified speckle frame from file
}
\examples{
# Plot frame number 57 from file you choose
speckle_frame(, 57) # no first argument here
# the same
speckle_frame(file.choose(), 57)
speckle_frame(data_file = file.choose(), 57)
}
