% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/speckle_frame.R
\name{speckle_frame}
\alias{speckle_frame}
\title{Plot selected speckle frame}
\usage{
speckle_frame(data_file = file.choose(), frame = 1)
}
\arguments{
\item{data_file}{A string.}

\item{frame}{An integer.}
}
\value{
512 x 512 matrix with given frame
}
\description{
Ploting specified speckle frame from file
}
\examples{
## Read frame number 57 from file you choose
# frame57 <- speckle_frame(, 57) # no first argument here
## the same
# frame57 <- speckle_frame(file.choose(), 57)
# frame57 <- speckle_frame(data_file = file.choose(), 57)
}
