% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{speckle_generator}
\alias{speckle_generator}
\title{Speckle Generator}
\usage{
speckle_generator(seeing, speckle_sigma, m1, m2, rho_x, rho_y, wind)
}
\arguments{
\item{seeing}{A number.}

\item{speckle_sigma}{A number.}

\item{m1}{A number.}

\item{m2}{A number.}

\item{rho_x}{A number.}

\item{rho_y}{A number.}

\item{wind}{A number.}
}
\value{
The vector of model speckle image.
}
\description{
Generate model 512 x 512 x 2 (bytes) speckle image of binary star
}
\section{Details}{

Details here.
Details here.
}

\examples{
## Generate speckle image of binary star with
## 7 parameters
# speckle_vector <- speckle_generator(seeing = 30, speckle_sigma = 1,
# m1 = 1000, m2 = 900, rho_x = 50, rho_y = 70, wind = 0)
# peckle_matrix <- matrix(speckle_vector, nrow = 512, ncol = 512)

## Plot result
# library(imager)
# plot(as.cimg(speckle_matrix))
}
