% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/speckle_limits.R
\name{speckle_limits}
\alias{speckle_limits}
\title{Detection limits from power spectrum}
\usage{
speckle_limits(ps)
}
\arguments{
\item{ps}{matrix with power spectrum.}
}
\value{
Vector with limits.
}
\description{
Calculate detection limits (dm for given rho) from empirical power spectrum
}
\examples{
obj_filename <- system.file("extdata", "ads15182_550_2_frames.dat", package = "specklestar")
pow_spec_diff <- speckle_ps_diff(obj_filename)
profile <- speckle_limits(ps = pow_spec_diff)
}
